backend/requirements.txt
fastapi
uvicorn
requests
langchain
matplotlib
python-dotenv 